removed empty component and moved the shared component and hook under a correct folder
