--- src/components/SavedCoffee.tsx
@@ -1,6 +1,0 @@
-function SavedCoffee() {
-  return;
-  <div>THIS IS SAVED COFFEE LIST</div>;
-}
-
-export default SavedCoffee;
